Fixing datafiles Perl API usage.
Seems that we never used the Bio::DB::Sam bindings properly. The value
we retrieved was total rubbish and we should have been using $header->n_targets().
The replaced method was getting an ArrayRef of reference sequence lengths. Not
what was required.

Also added in methods to allow users to specify a samtools binary if required
and force the use of the binary.