[ENSCORESW-604]. Raise information to the log table.
Confusion was caused over how the FASTA pipeline decides to redump
species. This particular cause was caused by the lack of an FTP
directory when it was run. So we will now tell the log table why
a species has been selected for a rerun outside of the normal reasons.