[ENSCORESW-603]. Module did not use full path.
Turns out this adaptation went wrong. The code was asssuming it could
access the contents of the data directory. Since the scripts CWD
was not the same as that of where the file was we failed to find the
file. Now we cat the filename onto the directory. This should work