[ENSCORESW-753]. Change permissions detection.
Due to some errors we are firming up the permissions code. We require files
to be read for user/group/other. They cannot be write for any designation.
We do not care about executable permissions (they shouldn't be but if they
are it is not the end of the world).