Update compute_closure.pl
After having a few issues with loading it was identified that there was an error in the closure loading script wrt the handling of links between aspects. I have modified the code in the between aspect and within aspect code to be more respectful and not cross the line.

Quick test load of the db had resulted in the correct output.