Provide -sort_memory for bam sorting
Bam sorting can be faster with enough RAM to hold the files to sort.

I also added a dummy option to the Aligner to be able to skip the execution
but still provide the executed command line (in case some files were
generated but the command line was not stored).