Exclude anything that does not match __\_%
We have had issues with linkout not finding some identifiers. So instead we filter out anything that does not look like a normal RefSeq identifier. We also removed the GROUP BY because it was not useful (a check on Ensembl human 89 showed no duplicates between primary accession and stable id).