sql2html: Use identify rather than file to extract PNG dimensions
The advantage of using identify (part of ImageMagick, or
GraphicsMagick with compatibility symlinks) is that it allows
extraction of dimensions from many more image formats than file (which
e.g. doesn't presently do so for SVG) and has the same output structure
for different formats (whereas that of file varies wildly between
formats - compare e.g. PNG and JPEG). The DISadvantage of using it is
that it introduces a dependency on ImageMagick/GraphicsMagick - but then
again, the latter is fairly common as well and it shares many
second-order dependencies with GraphViz (which we must have in order to
actually *produce* diagrams from which we subsequently extract
dimensions).

Side note: ImageMagick/GraphicsMagick come with Perl bindings so we might
want to switch to using them instead of invoking an external executable.
For now however let's keep the change simple.