sql2html: allow selection of diagram-file format from the command line
The new option is called -diagram_format and currently supports PNG and SVG.
The former is the default for backwards compatibility.

Note that a lot of the plumbing behind format selection, in particular having
to use a dispatch table to select the right $graph method, could be
dramatically simplified if we upgraded Bio::EnsEMBL::Hive::Utils::GraphViz
from the now-deprecated GraphViz to GraphViz2.