Remove defaults for datachecks
I think it's probably best to not have defaults for the registry and old server uri - the registry needs to match the 'db_host' parameter, which is passed to init_pipeline, so I think it makes sense to also pass the registry, to avoid inadvertently using a mismatched registry. And the old_server_uri needs to have the database included, which is easier to manage by requiring it to be passed every time.
I've set the history file to be undefined by default, since that's not an absolute requirement - the confluence SOP can include a file in which to store the histoey, by convention.