Parameterise queue names
The queue names on Codon are different, but if we switch
to a hard-coded value, that removes our ability to easily
run pipelines on Noah. During this transition period we
need to be able to run on both, because, for example,
we don't have ftp write access on Codon yet, so Noah
pipelines are still needed for production tasks.