Modify Protein Features pipeline for Codon cluster.
On Codon we have a centrally-managed version of InterProScan,
which is much easier for us to manage, but needs a little
bit of adaptation in the pipeline.

For a given instance of the pipeline, we want to run the
same version of interproscan; but the interproscan team
might update the 'current' symlink mid-pipeline.
So, resolve the 'current' link to a distinct version at
the start of the pipeline, and pass that version to
subsequent modules that need the path to the executable.

A corollary of the above point is that the specific version
does not need to be hard-coded in the config file.

Versions of the Interpro source dbs are also no longer
hard-coded in the conf, but are instead determined by
running interproscan.sh with no parameters, to get a list
of the dbs and their versions.