Update OLS code to use consistent output and scratch paths.
Use the default scratch space for the mart SQL file, and
delete the directory when done. Use standard 'pipeline_dir'
for pipeline-related files.
Also did a bit of tidying, pipeline-wide parameters were
being passed, with the same values, to several analyses;
at best this is redundant, but it could be a source of bugs
if an update happens in one place but not the other, so
removed them from the analyses.