Increase number of cores for InterProScan execution
While discussing tests of the software on Codon, InterProScan peeps
happened to mention that it uses up to 8 cores by default. We could
tell it to be happy with 4, by setting the -cpu parameter, and this should
not,apparently, affect performance significantly. But, we might as well
let the program have 8 cores, if that's the default - it should not make
much difference in terms of requesting LSF resources.