Make mysql dump module more flexible.
The mysql dumping module that was written for the RR dumps
(not currently used) was only able to dump in the RR directory
structure. There is a general override mechanism for the RR
dumps, that allows for more flexibility, but the nature of the
mysql dumps means that this is not inherited. So, if an
'output_dir' is provided, that now overrides the automatically
generated path.
The module also lacked the ability to dump 'multi' databases,
such as ensembl_metadata and ncbi_taxonomy - this can now be
done, with the addition of a 'db_url' parameter.
These two changes mean that the module can be used for mysql
dumps for the main Ensembl/EG sites - this is useful, because
the existing pipeline is complicated and non-standard (e.g.
it doesn't use registry files).