Additional mysql dumping parameters.
Allow for control of overwriting existing directories.
Enable a sub-directory to be given for use in the dump
path; this defaults to the division, but for GRCh37
dumps it would be set to 'grch37'.
Create CHECKSUMS file rather than md5sum; the latter is
better, but have to be consistent with rest of main site.