Ontology Load pipeline updates
To accommodate execution on Codon, inherit from the base
Production config. Split the datacheck execution into
advisory and critical sets; the comparison with the previous
release is almost guaranteed to fail, but will rarely be
something to worry about. If that test is critical, the
pipeline will almost always fail, and not create the dbcopy
jobs. So the comparison check is made advisory, and the
report is emailed. Renamed a couple of variables and
reformatted a bit, to be consistent with other pipelines.