Modify Protein Features pipeline, 'seg' execution.
We need a bit more flexibility in 'seg' execution, so that we
can always force it to run, even if there are already features
annotated. The use case is for human and mouse, where we might
have partial annotation, and we want to re-annotate everything.
The assumption of the pipeline now reflects usage, and treats
'seg' the same as the InterPro sources, in that it is always
run of features do not exist. By default, it is not run if
features exist, but this can be overriden by setting the
'run_seg' paramter to 1.