Datachecks added to file dumping pipelines.
The new pipeline to dump GFF3 for ENA submission requires a
pre-dump check, that all of the sequences have INSDC names or
synonyms. It's easiest to introduce this into the generic
dumping pipeline, and that turns out to be useful, because we
have a similar requirement for the Rapid Release dumps, which
depend on the display name. (Ultimately, we should add that
datacheck to the RR handover set, but this will do for now.)