Module and pipeline to update 'website_packed' status for Rapid Release.
"Packing" the data for a species, to enable the website display
to be quicker, is generally required whenever a database changes.
Packing can be a time-consuming process, however, so for Rapid
Release, where we have a two-week release cycle and lots of
databases that aren't changing from one release to the next,
we want to avoid redundant work.

A new field has been added to the metadata db to track the packed
status, and there is a script in the metadata repo to update this.
But it makes it easier for us to manage if we also have that
functionality in a module that we can add to pipelines, and a
pipeline that will apply to all species derived from a SpeciesFactory
module.