Dump FASTA DNA for renamed genomes.
The dumping pipeline was detecting new and updated assemblies and
redumping FASTA DNA files, and copying files from the previous
release. But it was missing renamed genomes, which need to be
redumped because the filename is different, even if the contents
are not. Renaming the file is complicated, it's easiest to just
redump in this case too.

Rather than augmenting the existing module to detect changes,
which is already complicated and essentially rehashes the work
of the report_genomes.pl script, introduce a new module that
parses the json report from that script. That file will already
have been created by an earlier process in the release cycle.