Remove existing assembly_converter data from directory.
When a genome was renamed, the existing data was not removed
from the directory, which caused a problem for the Web team.
Since the directory is populated with new data for every
release, delete the contents before starting.

Also, removed the functionality for processing verts and
non-verts separately - it made the above change simpler,
and we always run the script for everything at once anyway.