Remove pre-existing BLAT files.
Because the BLAT modules follow on from the FASTA DNA dumps,
the commits that ensured those happen for renamed species
will propagate to the BLAT file generation. An additional
fix is required in the BLAT indexes to remove existing files;
for the most part this is not necessary, because new files
with the same name overwrite old ones. However, when a species
is renamed or an assembly updated, we end up with a file for
both, which is unnecessary and confusing. (The exception is
human GRCh37, which we always need to keep as is.)