Update FTP Checker pipeline for Codon execution.
Set FTP Checker pipeline to inherit from the standard Base
config, to enable execution on the Codon cluster.
Also, a couple of improvements to the pipeline:
1. Drop TSV checking - we don't have clear expectations
of these files, so we end up with lots of spurious 'failures',
amongst which it's easy to miss important failures.
2. Rather than rely on the person running the pipeline to find
the results file and look at it, send an email with its
location, and attach file if it's not too big.