For MySQL file dumps, treat compara as 'multi' database.
We _can_ retrieve 'compara' dbs via the DbFactory, but it's
harder to get 'ancestal' dbs. Both of these are multispecies
dbs, so it makes sense to process them alongside all the
other dbs.
Also, it's easier, and more consistent with other dumps, if
we don't require division parameters for mart dumps, and
assume that we want them for all divisions, if not specified.