Mart db name trumps pan division
We have one database, the ontology_mart, which fits into
two categories being both a 'mart' and 'pan' (i.e. not
related to a specific division).

When processing the databases, we don't want the same db
appearing twice, and so we need to decide whether to treat
it as either 'mart' or 'pan' (and not as both).

In practice we tend to group all the marts together
(e.g. they have their own mirror server), so it best fits our
processes to give primacy to the mart status.