Pass division parameter to rename factory
The renaming needs to be restricted to a subset of dbs; otherwise
running the pipeline for bacteria leads to retrieval of db names
for all divisions from the metadata db, which then fails for the
non-bacterial dbs. Rather than ignoring those, which could lead
to us overlooking true discrepancies in the expected dbs, filter
based on division. Our SOP is that we will always be passing this
parameter in any case, to get the relevant marts and comparas.

Also, renamed the parameter from 'division' to 'multi_division',
to avoid confusion with the SpeciesFactory 'division' parameter,
since the usage/behaviour is different here.