AlphaFold pipeline: integrate map file creation
The AlphaFold pipeline currently uses mapping files to associate proteins in a
species with the prediction data. This commit changes the pipeline to create
these mapping files in a first step.
The pipeline uses scratch_large_dir, which is created at the beginning and
deleted when the pipelines has finished.
Includes a dependency on Gzip::Faster. It is 8 - 10x faster than the
default IO::Compress::Gzip.