Alphafold pipeline rework, now based on Uniparc
The Alphafold pipeline has been reworked. Instead of the .pdb.gz per species, we
now use the accession_ids.csv which contains all accessions.
Additionally, for species not in the GIFTS DB, we use the Uniparc accessions we
have in our core DBs to map to the Alphafold data. The Uniparc accession
mappings are of higher quality than the other mappings available.
To match Uniparc, Uniprot and Alphafold, we now build temporary DBs in the
LevelDB format. For performance reasons, we do this in-memory if possible.