Read data for GIFTS from file, not REST endpoint
So far, we used the REST API endpoint to retrieve GIFTS data. This turned out to
be error prone. Fetching data this way takes many hours. If the process is
interrupted, the error would be ignored and Uniparc data would be used.
Now we use dump files instead. They take seconds to create and to ingest.